refactor(ui): cleanup PR #7 — extract EventFlows, remove MenuBuilder aliases, unify history
- Extract EventMenu inner classes (MainFlow/EventsFlow/EventFlow) to EventFlows.java
- Remove 12 alias methods from MenuBuilder, keep Text/Key as canonical naming
- Unify canGoBack(boolean) to canGoBack() — always checks both history stacks
- Remove 3 redundant clearHistory() calls (already handled by session.clear())
- Deprecate setDraft(Object) in favor of setDraft(Class, T), migrate 2 callers
- Update MenuBuilderTest to match unified canGoBack() behavior

Net: -518 lines removed, +44 lines changed, EventFlows.java added (451 lines)
