+++ src/main/java/org/xcore/plugin/ui/menu/AuditHistoryFlows.java
@@ -110,7 +110,7 @@ final class AuditHistoryFlows {
             }
 
             List<MenuButton> navRow = new ArrayList<>();
-            if (session.canGoBack(true)) {
+            if (session.canGoBack()) {
                 navRow.add(MenuButton.of(local.t("back"), ACTION_BACK));
             }
             navRow.add(MenuButton.of(local.t("close"), ACTION_CLOSE));
@@ -203,7 +203,7 @@ final class AuditHistoryFlows {
 
             List<List<MenuButton>> rows = new ArrayList<>();
             List<MenuButton> navRow = new ArrayList<>();
-            if (session.canGoBack(true)) {
+            if (session.canGoBack()) {
                 navRow.add(MenuButton.of(local.t("back"), ACTION_BACK));
             }
             navRow.add(MenuButton.of(local.t("close"), ACTION_CLOSE));
