+++ src/main/java/org/xcore/plugin/ui/menu/EventDraftFlows.java
@@ -348,7 +348,7 @@ final class EventDraftFlows {
             }
 
             List<MenuButton> navigation = new ArrayList<>();
-            if (session.canGoBack(true)) {
+            if (session.canGoBack()) {
                 navigation.add(MenuButton.of(session.locale().t("back"), ACTION_BACK));
             }
             navigation.add(MenuButton.of(session.locale().t("close"), ACTION_CLOSE));
