+++ src/main/java/org/xcore/plugin/ui/menu/HelpFlows.java
@@ -99,7 +99,7 @@ final class HelpFlows {
             }
 
             List<MenuButton> navigation = new ArrayList<>();
-            if (session.canGoBack(true)) {
+            if (session.canGoBack()) {
                 navigation.add(MenuButton.of(local.t("back"), ACTION_BACK));
             }
             navigation.add(MenuButton.of(local.t("close"), ACTION_CLOSE));
@@ -190,7 +190,7 @@ final class HelpFlows {
             rows.add(List.of(MenuButton.of(local.t("help-back"), ACTION_BACK)));
 
             List<MenuButton> navigation = new ArrayList<>();
-            if (session.canGoBack(true)) {
+            if (session.canGoBack()) {
                 navigation.add(MenuButton.of(local.t("back"), ACTION_BACK));
             }
             navigation.add(MenuButton.of(local.t("close"), ACTION_CLOSE));
