+++ src/main/java/org/xcore/plugin/ui/menu/PlayerProfileFlows.java
@@ -116,7 +116,7 @@ final class PlayerProfileFlows {
             rows.add(List.of(MenuButton.of(local.t("player-menu-players"), ACTION_PLAYERS)));
 
             List<MenuButton> navigation = new ArrayList<>();
-            if (session.canGoBack(true)) {
+            if (session.canGoBack()) {
                 navigation.add(MenuButton.of(local.t("back"), ACTION_BACK));
             }
             navigation.add(MenuButton.of(local.t("close"), ACTION_CLOSE));
@@ -267,7 +267,7 @@ final class PlayerProfileFlows {
             }
 
             List<MenuButton> navigation = new ArrayList<>();
-            if (session.canGoBack(true)) {
+            if (session.canGoBack()) {
                 navigation.add(MenuButton.of(local.t("back"), ACTION_BACK));
             }
             navigation.add(MenuButton.of(local.t("close"), ACTION_CLOSE));
