+++ src/main/java/org/xcore/plugin/ui/menu/TopMenu.java
@@ -175,7 +175,7 @@ public class TopMenu extends Menu {
             }
 
             List<MenuButton> bottomNav = new ArrayList<>();
-            if (session.canGoBack(true)) {
+            if (session.canGoBack()) {
                 bottomNav.add(MenuButton.of(local.t("back"), ACTION_BACK));
             }
             bottomNav.add(MenuButton.of(local.t("close"), ACTION_CLOSE));
@@ -296,7 +296,7 @@ public class TopMenu extends Menu {
             ));
 
             List<MenuButton> navRow = new ArrayList<>();
-            if (session.canGoBack(true)) {
+            if (session.canGoBack()) {
                 navRow.add(MenuButton.of(local.t("back"), ACTION_BACK));
             }
             navRow.add(MenuButton.of(local.t("close"), ACTION_CLOSE));
